# docs/swagger.yaml
basePath: /api/
definitions:
  api.ParamSearch:
    properties:
      filter:
        additionalProperties: true
        type: object
      keyword:
        additionalProperties:
          type: string
        type: object
      limit:
        type: integer
      skip:
        type: integer
      sort:
        additionalProperties:
          type: integer
        type: object
    type: object
...
  api.ReplyData-int64:
    properties:
      data:
        type: integer
      error:
        type: string
    type: object
...
  api.ReplyData-types_Link:
    properties:
      data:
        $ref: '#/definitions/types.Link'
      error:
        type: string
    type: object
  api.ReplyData-types_Product:
    properties:
      data:
        $ref: '#/definitions/types.Product'
      error:
        type: string
    type: object
...
  api.ReplyList-types_Link:
    properties:
      data:
        items:
          $ref: '#/definitions/types.Link'
        type: array
      error:
        type: string
      total:
        type: integer
    type: object
  api.ReplyList-types_Product:
    properties:
      data:
        items:
          $ref: '#/definitions/types.Product'
        type: array
      error:
        type: string
      total:
        type: integer
    type: object
...
  types.Link:
    properties:
      created:
        description: 创建时间
        type: string
      desc:
        description: 说明
        type: string
      disabled:
        type: boolean
      heartbeat:
        description: 心跳包
        type: string
      id:
        description: ID
        type: string
      name:
        description: 名称
        type: string
      poller_interval:
        description: 采集间隔
        type: integer
      poller_period:
        description: 采集周期
        type: integer
      protocol_name:
        description: 协议 rtu tcp parallel-tcp
        type: string
      protocol_options:
        description: 协议参数
        type: string
      remote:
        description: 远程地址
        type: string
      server_id:
        description: 服务器ID
        type: string
    type: object
  types.Mapper:
    properties:
      addr:
        description: 地址
        type: integer
      code:
        description: 指令
        type: integer
      points:
        description: 数据点
        items:
          $ref: '#/definitions/types.Point'
        type: array
      size:
        description: 长度
        type: integer
    type: object
  types.Point:
    properties:
      be:
        description: 大端模式
        type: boolean
      name:
        description: 名称
        type: string
      offset:
        description: 偏移
        type: integer
      rate:
        description: 倍率
        type: number
      type:
        description: 类型
        type: string
    type: object
  types.Product:
    properties:
      created:
        description: 创建时间
        type: string
      desc:
        description: 说明
        type: string
      id:
        type: string
      mappers:
        description: model.Product `xorm:"extends"`
        items:
          $ref: '#/definitions/types.Mapper'
        type: array
      name:
        description: 名称
        type: string
    type: object
...
info:
  contact: {}
  description: API文档
  title: 物联大师网关接口文档
  version: 1.0 版本
paths:
...
  /link/{id}:
    get:
      consumes:
      - application/json
      description: 获取连接
      parameters:
      - description: 连接ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ReplyData-types_Link'
      summary: 获取连接
      tags:
      - link
    post:
      consumes:
      - application/json
      description: 修改连接
      parameters:
      - description: 连接ID
        in: path
        name: id
        required: true
        type: integer
      - description: 连接信息
        in: body
        name: link
        required: true
        schema:
          $ref: '#/definitions/types.Link'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ReplyData-types_Link'
      summary: 修改连接
      tags:
      - link
  /link/{id}/delete:
    get:
      consumes:
      - application/json
      description: 删除连接
      parameters:
      - description: 连接ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ReplyData-types_Link'
      summary: 删除连接
      tags:
      - link
  /link/{id}/disable:
    get:
      consumes:
      - application/json
      description: 禁用连接
      parameters:
      - description: 连接ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ReplyData-types_Link'
      summary: 禁用连接
      tags:
      - link
  /link/{id}/enable:
    get:
      consumes:
      - application/json
      description: 启用连接
      parameters:
      - description: 连接ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ReplyData-types_Link'
      summary: 启用连接
      tags:
      - link
  /link/count:
    post:
      consumes:
      - application/json
      description: 查询连接数量
      parameters:
      - description: 查询参数
        in: body
        name: search
        required: true
        schema:
          $ref: '#/definitions/api.ParamSearch'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ReplyData-int64'
      summary: 查询连接数量
      tags:
      - link
...
  /link/import:
    post:
      consumes:
      - multipart/form-data
      description: 导入连接
      parameters:
      - description: 压缩包
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ReplyData-int64'
      summary: 导入连接
      tags:
      - product
  /link/list:
    get:
      consumes:
      - application/json
      description: 查询连接
      parameters:
      - in: query
        name: limit
        type: integer
      - in: query
        name: skip
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ReplyList-types_Link'
      summary: 查询连接
      tags:
      - link
  /link/search:
    post:
      consumes:
      - application/json
      description: 查询连接
      parameters:
      - description: 查询参数
        in: body
        name: search
        required: true
        schema:
          $ref: '#/definitions/api.ParamSearch'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ReplyList-types_Link'
      summary: 查询连接
      tags:
      - link
  /product/{id}:
    get:
      consumes:
      - application/json
      description: 获取产品
      parameters:
      - description: 产品ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ReplyData-types_Product'
      summary: 获取产品
      tags:
      - product
    post:
      consumes:
      - application/json
      description: 修改产品
      parameters:
      - description: 产品ID
        in: path
        name: id
        required: true
        type: integer
      - description: 产品信息
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/types.Product'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ReplyData-types_Product'
      summary: 修改产品
      tags:
      - product
  /product/{id}/delete:
    get:
      consumes:
      - application/json
      description: 删除产品
      parameters:
      - description: 产品ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ReplyData-types_Product'
      summary: 删除产品
      tags:
      - product
  /product/count:
    post:
      consumes:
      - application/json
      description: 查询产品数量
      parameters:
      - description: 查询参数
        in: body
        name: search
        required: true
        schema:
          $ref: '#/definitions/api.ParamSearch'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ReplyData-int64'
      summary: 查询产品数量
      tags:
      - product
  /product/create:
    post:
      consumes:
      - application/json
      description: 创建产品
      parameters:
      - description: 产品信息
        in: body
        name: search
        required: true
        schema:
          $ref: '#/definitions/types.Product'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ReplyData-types_Product'
      summary: 创建产品
      tags:
      - product
...
  /product/import:
    post:
      consumes:
      - multipart/form-data
      description: 导入产品
      parameters:
      - description: 压缩包
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ReplyData-int64'
      summary: 导入产品
      tags:
      - product
  /product/list:
    get:
      consumes:
      - application/json
      description: 查询产品
      parameters:
      - in: query
        name: limit
        type: integer
      - in: query
        name: skip
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ReplyList-types_Product'
      summary: 查询产品
      tags:
      - product
  /product/search:
    post:
      consumes:
      - application/json
      description: 查询产品
      parameters:
      - description: 查询参数
        in: body
        name: search
        required: true
        schema:
          $ref: '#/definitions/api.ParamSearch'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ReplyList-types_Product'
      summary: 查询产品
      tags:
      - product
...
swagger: "2.0"